copied in upstream packaging scripts

+++ install.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+# run build_everything.sh first
 
 cd `dirname $0`
 
@@ -16,9 +17,16 @@ else
 fi
 
 if [ "$1" = "-f" ] ; then
-    echo Removing existing installation...
+    echo Removing existing installation at $INSTALLDIR
+    echo -n 'Continue (y/N)? '
+    read cont
+    if [ "$cont" != "y" ] ; then
+        exit 0
+    fi
     rm -fv *.zip || exit 1
-    rm -r "$INSTALLDIR/"* || exit 1
+    set -x
+    sudo rm -r "$INSTALLDIR/"* || exit 1
+    set +x
 fi
 
 ./makezip.sh || exit 1
